requirements.txt
streamlit
moviepy
openai-whisper
torch

